Move helpers to shared support
