--- sequel/benchmarks/support/helpers.rb
@@ -1,3 +1,0 @@
-def db_adapter
-  ENV['DATABASE_URL'].split(":")[0]
-end
